requirements.txt
arxiv>=2.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyMuPDF>=1.24.0
rich>=13.0.0
